chore: add stop env

+++ crates/rspack_util/src/debug_tool.rs
@@ -1,7 +1,13 @@
 use signal_hook::{consts::signal::SIGUSR2, iterator::Signals};
+pub fn should_stop() -> bool {
+  return std::env::var("RSPACK_DEBUG_STOP").is_ok();
+}
 // wait for user input to continue, useful for debugging and profiling purposes
 // inspired by https://github.com/rust-lang/rust-analyzer/blob/661e7d2ac245f4fca099d982544b3c5408322867/crates/rust-analyzer/src/bin/main.rs#L97
 pub fn wait_for_enter(phase: &str) {
+  if !should_stop() {
+    return;
+  }
   println!(
     "Waiting in phase {} with pid={}, Press Enter to continue.",
     phase,
@@ -12,6 +18,9 @@ pub fn wait_for_enter(phase: &str) {
 }
 // wait for signal to continue
 pub fn wait_for_signal(phase: &str) {
+  if !should_stop() {
+    return;
+  }
   println!(
     "Waiting in phase {}, run `kill -SIGUSR2 {}` to continue.",
     phase,
